move routes to extensions
