+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/routes/definitions/not-found.js
